whisper_test/requirements.txt
accelerate==1.3.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.12
aiosignal==1.3.2
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
datasets==3.2.0
dill==0.3.8
evaluate==0.4.3
filelock==3.13.1
frozenlist==1.5.0
fsspec==2024.6.1
huggingface-hub==0.28.1
idna==3.10
Jinja2==3.1.4
jiwer==3.1.0
llvmlite==0.44.0
MarkupSafe==2.1.5
more-itertools==10.6.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.3
numba==0.61.0
numpy==2.1.2
openai-whisper @ git+https://github.com/openai/whisper.git@517a43ecd132a2089d85f4ebc044728a71d49f6e
packaging==24.2
pandas==2.2.3
pillow==11.0.0
propcache==0.2.1
psutil==6.1.1
pyarrow==19.0.0
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
RapidFuzz==3.12.1
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
setuptools==75.8.0
six==1.17.0
sympy==1.13.1
tiktoken==0.8.0
tokenizers==0.21.0
torch==2.6.0+cpu
torchaudio==2.6.0+cpu
torchvision==0.21.0+cpu
tqdm==4.67.1
transformers==4.48.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wheel==0.45.1
xxhash==3.5.0
yarl==1.18.3
